[refactor] 공통정보 삭제도 마저 분리
